setting the top level nav to open and using localstorge to save the open/closed selection
